Mac-Uebersetzung: Double und CGFloat nicht mischen
Der erste Lauf der ganzen App-Schicht auf dem Mac fand genau einen
Fehler: In der Uebergabe-Strecke wurde eine Laenge (CGFloat) mit einem
Anteil (Double) verrechnet, und Swift kann nicht entscheiden, welche
Division gemeint ist. Der vermutete fehlende Import war keiner.

Dazu Protokoll 67, Lexikon (17 App-Begriffe), Plan und Schritte-Blatt.

+++ ipad/Visbox.swiftpm/Verbindung/Uebergabe.swift
@@ -48,9 +48,9 @@ struct Uebergabestrecke: View {
                         .frame(height: 1)
                     Rectangle()
                         .fill(Uebergabestrecke.gruen)
-                        .frame(width: weg * ort + Uebergabestrecke.marke / 2, height: 2)
+                        .frame(width: weg * CGFloat(ort) + Uebergabestrecke.marke / 2, height: 2)
                     markeAnsicht
-                        .offset(x: weg * markenort)
+                        .offset(x: weg * CGFloat(markenort))
                 }
                 .frame(height: Uebergabestrecke.marke)
             }
